Move `FileDescriptors` under `tools` package
