+++ codegen/src/main/java/io/spine/tools/FileDescriptors.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.gradle.compiler.util;
+package io.spine.tools;
 
 import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
