+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/annotation/ProtoAnnotatorPlugin.java
@@ -26,7 +26,7 @@ import io.spine.annotation.Experimental;
 import io.spine.annotation.Internal;
 import io.spine.annotation.SPI;
 import io.spine.gradle.SpinePlugin;
-import io.spine.gradle.compiler.util.FileDescriptors;
+import io.spine.tools.FileDescriptors;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
@@ -46,7 +46,6 @@ import static io.spine.gradle.compiler.Extension.getMainGenProtoDir;
 import static io.spine.gradle.compiler.Extension.getTestDescriptorSetPath;
 import static io.spine.gradle.compiler.Extension.getTestGenGrpcDir;
 import static io.spine.gradle.compiler.Extension.getTestGenProtoDir;
-import static io.spine.gradle.compiler.util.FileDescriptors.isNotGoogleProto;
 import static io.spine.option.OptionsProto.beta;
 import static io.spine.option.OptionsProto.betaAll;
 import static io.spine.option.OptionsProto.betaType;
@@ -60,6 +59,7 @@ import static io.spine.option.OptionsProto.sPI;
 import static io.spine.option.OptionsProto.sPIAll;
 import static io.spine.option.OptionsProto.sPIService;
 import static io.spine.option.OptionsProto.sPIType;
+import static io.spine.tools.FileDescriptors.isNotGoogleProto;
 import static org.slf4j.LoggerFactory.getLogger;
 
 /**
