+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/lookup/proto/ProtoToJavaMapperPlugin.java
@@ -21,8 +21,8 @@ package io.spine.gradle.compiler.lookup.proto;
 
 import com.google.protobuf.DescriptorProtos.FileDescriptorProto;
 import io.spine.gradle.SpinePlugin;
-import io.spine.gradle.compiler.util.FileDescriptors;
 import io.spine.gradle.compiler.util.PropertiesWriter;
+import io.spine.tools.FileDescriptors;
 import io.spine.type.KnownTypes;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
@@ -44,7 +44,7 @@ import static io.spine.gradle.compiler.Extension.getMainDescriptorSetPath;
 import static io.spine.gradle.compiler.Extension.getMainTargetGenResourcesDir;
 import static io.spine.gradle.compiler.Extension.getTestDescriptorSetPath;
 import static io.spine.gradle.compiler.Extension.getTestTargetGenResourcesDir;
-import static io.spine.gradle.compiler.util.FileDescriptors.isNotGoogleProto;
+import static io.spine.tools.FileDescriptors.isNotGoogleProto;
 
 /**
  * Plugin which maps all Protobuf types to the corresponding Java classes.
