+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/validate/MetadataAssembler.java
@@ -24,7 +24,7 @@ import com.google.common.base.Predicate;
 import com.google.protobuf.DescriptorProtos.DescriptorProto;
 import com.google.protobuf.DescriptorProtos.FileDescriptorProto;
 import io.spine.gradle.compiler.message.MessageTypeCache;
-import io.spine.gradle.compiler.util.FileDescriptors;
+import io.spine.tools.FileDescriptors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
